+++ R/define_S4_generic_RCM.R
@@ -28,6 +28,10 @@
 #' @param ESS If \code{comp_like = "multinomial"}, a numeric vector of length two to cap the maximum effective samples size of the age and length compositions,
 #' respectively, for the multinomial likelihood function. The effective sample size of an age or length composition sample is the minimum of ESS or the number of observations
 #' (sum across columns). For more flexibility, set ESS to be very large and alter the age and length arrays as needed.
+#' @param prior A named list (R0, h, M, and q) to provide the mean and standard deviations of prior distributions for those parameters. R0 and M priors
+#' lognormal (mean in normal space, SD in lognormal space). Beverton-Holt steepness uses a beta prior, while survey q and Ricker steepness use normal priors.
+#' For survey q, provide a matrix for nsurvey rows and 2 columns (for mean and SD). For all others, provide a length-2 vector for the mean and SD.
+#' See vignette for full description.
 #' @param max_F The maximum F for any fleet in the scoping model (higher F's in the model are penalized in the objective function). See also `drop_highF`.
 #' @param cores Integer for the number of CPU cores for the stock reduction analysis.
 #' @param integrate Logical, whether to treat recruitment deviations as penalized parameters in the likelihood (FALSE) or random effects to be marginalized out of the likelihood (TRUE).
@@ -175,11 +179,11 @@ setGeneric("RCM", function(OM, data, ...) standardGeneric("RCM"))
 #' @export
 setMethod("RCM", signature(OM = "OM", data = "list"),
           function(OM, data, condition = c("catch", "catch2", "effort"), selectivity = "logistic", s_selectivity = NULL, LWT = list(),
-                   comp_like = c("multinomial", "lognormal"), ESS = c(30, 30),
+                   comp_like = c("multinomial", "lognormal"), ESS = c(30, 30), prior = list(),
                    max_F = 3, cores = 1L, integrate = FALSE, mean_fit = FALSE, drop_nonconv = FALSE,
                    drop_highF = FALSE, control = list(iter.max = 2e+05, eval.max = 4e+05), ...) {
             RCM_int(OM = OM, data = data, condition = condition, selectivity = selectivity, s_selectivity = s_selectivity, LWT = LWT,
-                    comp_like = comp_like, ESS = ESS, max_F = max_F, cores = cores, integrate = integrate, mean_fit = mean_fit,
+                    comp_like = comp_like, ESS = ESS, prior = prior, max_F = max_F, cores = cores, integrate = integrate, mean_fit = mean_fit,
                     drop_nonconv = drop_nonconv, drop_highF = drop_highF, control = control, ...)
           })
 
@@ -188,7 +192,7 @@ setMethod("RCM", signature(OM = "OM", data = "list"),
 #' @export
 setMethod("RCM", signature(OM = "OM", data = "Data"),
           function(OM, data, condition = c("catch", "catch2", "effort"), selectivity = "logistic", s_selectivity = NULL, LWT = list(),
-                   comp_like = c("multinomial", "lognormal"), ESS = c(30, 30),
+                   comp_like = c("multinomial", "lognormal"), ESS = c(30, 30), prior = list(),
                    max_F = 3, cores = 1L, integrate = FALSE, mean_fit = FALSE, drop_nonconv = FALSE,
                    drop_highF = FALSE, control = list(iter.max = 2e+05, eval.max = 4e+05), ...) {
 
@@ -257,7 +261,7 @@ setMethod("RCM", signature(OM = "OM", data = "Data"),
 
             ####### Run RCM
             output <- RCM_int(OM = OM, data = data_list, condition = condition, selectivity = selectivity, s_selectivity = Ind$s_sel, LWT = LWT,
-                              comp_like = comp_like, ESS = ESS, max_F = max_F, cores = cores, integrate = integrate, mean_fit = mean_fit,
+                              comp_like = comp_like, ESS = ESS, prior = prior, max_F = max_F, cores = cores, integrate = integrate, mean_fit = mean_fit,
                               drop_nonconv = drop_nonconv, drop_highF = drop_highF, control = control,
                               OMeff = extra_args$OMeff, s_vul_par = extra_args$s_vul_par, map_s_vul_par = extra_args$map_s_vul_par, ...)