+++ inst/include/ns/ns_RCM.hpp
@@ -177,7 +177,7 @@ array<Type> calc_vul_sur(matrix<Type> vul_par, vector<int> vul_type, matrix<Type
 // Calculates analytical solution of catchability when conditioned on catch and
 // index is lognormally distributed.
 template<class Type>
-Type calc_q(matrix<Type> I_y, matrix<Type> B_y, int sur, int ff, matrix<Type> &Ipred, vector<int> abs_I) {
+Type calc_q(matrix<Type> I_y, matrix<Type> B_y, int sur, int ff, matrix<Type> &Ipred, vector<int> abs_I, int n_y) {
   Type q;
   if(abs_I(sur)) {
     q = Type(1);
@@ -185,7 +185,7 @@ Type calc_q(matrix<Type> I_y, matrix<Type> B_y, int sur, int ff, matrix<Type> &I
     Type num = 0.;
     Type n_y = 0.;
 
-    for(int y=0;y<I_y.rows();y++) {
+    for(int y=0;y<n_y;y++) {
       if(!R_IsNA(asDouble(I_y(y,sur))) && I_y(y,sur)>0) {
         num += log(I_y(y,sur)/B_y(y,ff));
         n_y += 1.;
@@ -193,7 +193,7 @@ Type calc_q(matrix<Type> I_y, matrix<Type> B_y, int sur, int ff, matrix<Type> &I
     }
     q = exp(num/n_y);
   }
-  for(int y=0;y<I_y.rows();y++) Ipred(y,sur) = q * B_y(y,ff);
+  for(int y=0;y<n_y;y++) Ipred(y,sur) = q * B_y(y,ff);
   return q;
 }
 
@@ -222,6 +222,27 @@ Type comp_lognorm(array<Type> obs, array<Type> pred, matrix<Type> N, matrix<Type
   return log_like;
 }
 
-
+template<class Type>
+Type RCM_prior(matrix<int> use_prior, matrix<Type> prior_dist, Type R0, Type h, int SR_type, Type log_M, vector<Type> q) {
+  Type prior = 0;
+  if(use_prior(0)) { // Prior for R0 - normal on log_R0, log Jacobian transform = zero
+    prior += dnorm_(log(R0), prior_dist(0,0), prior_dist(0,1), true);
+  }
+  if(use_prior(1)) { // Prior for h
+    if(SR_type) { // Beverton-Holt - beta on y = (h - 0.2)/0.8 with log Jacobian transform of inverse logit fn
+      Type y = (h - 0.2)/0.8;
+      prior += dbeta_(y, prior_dist(1,0), prior_dist(1,1), true) - log(y - y * y); 
+    } else { // Ricker - normal on h with log Jacobian transform
+      prior += dnorm_(h, prior_dist(1,0), prior_dist(1,1), true) - log(h - 0.2);
+    }
+  }
+  if(use_prior(2)) { // Prior for constant M - normal on log_M
+    prior += dnorm_(log_M, prior_dist(2,0), prior_dist(2,1), true);
+  }
+  for(int i=3;i<use_prior.size();i++) { // Prior for q - normal
+    if(use_prior(i)) prior += dnorm_(q(i-3), prior_dist(i,0), prior_dist(i,1), true);
+  }
+  return prior;
+}
 
 }