+++ NEWS.md
@@ -9,5 +9,5 @@ The latest release of the SAMtool package is available on [CRAN](https://CRAN.R-
 - The function for the OM conditioning model is now re-named to `RCM` (Rapid Conditioning Model). 
 - The age structure of the model now starts at age 0, following the change in the MSEtool OM. The dimension associated with age in matrices and arrays need to be of length `maxage + 1` which corresponds to ages 0 to maxage.
 - When estimating fleet F in the model (`condition = "catch"`), the likelihood for the catch can now have a user-defined standard deviation indicated in `data$C_sd` (year and fleet specific, the previous default was 0.01 was built-in for all catches).
-- For generating length comps, variability in length-at-age can now vary by age. Specify the length-at-age standard deviation in `OM@cpars$LatASD`.
-
+- For generating length comps, variability in length-at-age can now be age-specific. Specify the length-at-age standard deviation in `OM@cpars$LatASD`.
+- Priors for log_R0, steepness, log_M (age and time constant), and survey q can now be specified.